Changed to double in heuristic

+++ algorithm/heuristic/heuristic.c
@@ -56,7 +56,7 @@ Result get_result(BoardState *board_state, bool is_finished)
 
 BoardScore score_board(BoardState *board_state, uint8_t depth, bool is_finished)
 {
-    int score = 0;
+    double score = 0;
     Result result = get_result(board_state, is_finished);
 
     // Material counting
@@ -73,64 +73,6 @@ BoardScore score_board(BoardState *board_state, uint8_t depth, bool is_finished)
 
     // Square control scoring
     score += get_square_control(board_state);
-
-    // // In score_board():
-    // score += __builtin_popcountll(board_state->white_pieces & CENTER_SQUARES_MASK) * 20;
-    // score -= __builtin_popcountll(board_state->black_pieces & CENTER_SQUARES_MASK) * 20;
-    // score += __builtin_popcountll(board_state->white_pieces & EXTENDED_CENTER_MASK) * 10;
-    // score -= __builtin_popcountll(board_state->black_pieces & EXTENDED_CENTER_MASK) * 10;
-
-    // // Bonus for protected pieces
-    // score += __builtin_popcountll(board_state->white_attack & board_state->board.white_pieces.queens) * 10;
-    // score += __builtin_popcountll(board_state->white_attack & board_state->board.white_pieces.rooks) * 5;
-    // score += __builtin_popcountll(board_state->white_attack & board_state->board.white_pieces.bishops) * 5;
-    // score += __builtin_popcountll(board_state->white_attack & board_state->board.white_pieces.knights) * 5;
-    // score += __builtin_popcountll(board_state->white_attack & board_state->board.white_pieces.pawns) * 2;
-
-    // score -= __builtin_popcountll(board_state->black_attack & board_state->board.black_pieces.queens) * 10;
-    // score -= __builtin_popcountll(board_state->black_attack & board_state->board.black_pieces.rooks) * 5;
-    // score -= __builtin_popcountll(board_state->black_attack & board_state->board.black_pieces.bishops) * 5;
-    // score -= __builtin_popcountll(board_state->black_attack & board_state->board.black_pieces.knights) * 5;
-    // score -= __builtin_popcountll(board_state->black_attack & board_state->board.black_pieces.pawns) * 2;
-
-    // // Bonus for attacking the king
-    // score += __builtin_popcountll(board_state->white_attack & board_state->board.black_pieces.king) * 20;
-    // score -= __builtin_popcountll(board_state->black_attack & board_state->board.white_pieces.king) * 20;
-    // // Bonus for having the move
-    // if (board_state->board.side_to_move == WHITE)
-    //     score += 5;
-    // else
-    //     score -= 5;
-
-    // // Bonus for rooks on open files
-    // uint64_t white_rooks = board_state->board.white_pieces.rooks;
-    // uint64_t black_rooks = board_state->board.black_pieces.rooks;
-    // uint64_t white_open_files = 0;
-    // uint64_t black_open_files = 0;
-    // for (int i = 0; i < 8; i++)
-    // {
-    //     uint64_t file = 0x0101010101010101ULL << i;
-    //     if ((white_rooks & file) == 0)
-    //         white_open_files |= file;
-    //     if ((black_rooks & file) == 0)
-    //         black_open_files |= file;
-    // }
-    // score += __builtin_popcountll(white_open_files & board_state->white_attack & board_state->board.white_pieces.rooks) * 10;
-    // score -= __builtin_popcountll(black_open_files & board_state->black_attack & board_state->board.black_pieces.rooks) * 10;
-
-    // // Bonus for connected rooks
-    // if ((board_state->board.white_pieces.rooks & board_state->white_attacks.rooks) != 0)
-    //     score += 10;
-    // if ((board_state->board.black_pieces.rooks & board_state->black_attacks.rooks) != 0)
-    //     score -= 10;
-
-    // // King safety
-    // // - Bonus for pawn shelter in front of the king
-    // // - Penalty for open files or half-open file next to king
-    // // Bonus for number of moves for sliding pieces
-    // // Taxicab (Distance from queen to king)
-    // // Pawn storm
-    // // Attacking squares around the king
 
     if (board_state->board.side_to_move == BLACK)
         score = -score;
